Created new solution "Debugging"
Created basic program that returns a specified ammount of the shortest names in a list of strings.
Created a logic error in order to use the debugging tools and simulate the bug hunting.

Learned the debugging basics. (How to use and look up breakpoints)
Learned about Locals and Autos.
Learned about debugging, creating copies of lists and how to hunt bugs.
Learned about the call stack, throwing errors and defensive programming.

Now on to Inheritance and more about Object Oriented Programming.